Globus project uploaded!!!
Hey guys,

Sorry about the wait. The Globus welcome screen, including all 3
fragments we planned on. There aren't as many comments as I'd like, but
I think there are enough for David and Taylor to get started. Please let
me know if it's confusing in any way, and I'll walk you through it.
I also made a new folder inside of Documents to store our iterations of
the android project. I wasn't sure if there was somewhere else it was
supposed to go, so if that already exists, feel free to change the
location.
-Jesse