addEvent in Calendar_Screen updated
Hey, just finished initial coding on the addEvent function. Had to
update the eventItem a little bit while working on this, so that's been
done too. I had to ad a month, date and year field to it. If we want
events that last multiple days, we will probably need to add an end
month date and year as well, though.