calendar event fragment and google drive project fragments
the event fragment is almost all the way done. The xml hates me and I'm
done with it for now. There's an error in the google_drive stuff
somewhere in the xml. I don't know what it is, but shouldn't be terrible
to fix. maybe try looking at jim wards project?