Final commit for the night.
Alright, after finishing up what I was doing, I took Kelsey's code and
got it up and running. Here is the total list of what's changed here.

-The Whiteboard
1. Wrote base code for Whiteboard, allowing it to  read in
events/messages from a list of "whiteboardItems" and display these items
on screen.
1.1 added two test items to appear automatically so you can see how they
work.
2. Added funcionality to the "add message" button. It will now prompt a
dialog (formatted via XML file) which will ask user for all necessary
info, build a message and automatically update the whiteboard.
- Calendar view
1. Changed Kelsey's code into something similar to that of what I did
with the "create message" button in the white board. Turned it into a
dialog. At this point, it doesn't do anything, but it does appear when
the button is pressed.

Besides this, I can't remember if I did anything else or not.