Final revision pre-mobile project
Updates:
- Added new custom adapter to group select screen
- Added on click listener for items in the list
- Fixed Log in credential check bug
-- Needed to change String comparison from "=="
to ".equals()"
- Compressed whiteboard icons
-- Were causing phone to run out of memory
- Realized you cannot add events to a calendar view
-- Questioned why
-- searched for substitute. Found one, but could not
import it correctly. We can always this try later.
- Updated whiteboard event display to show name of
events/messages inside description.
Note: I'm sorry I couldn't get the calendar working tonight. I was
looking up how to add events to it, when I learned that you can't. It's
really just there to show what dates are when and to stem to different
things. Not to indicate any date specific events. However, I did find a
custom calendar that could work for us. Sadly, I couldn't find a way to
import it this evening, so we won't have it for 4730, but we can always
use it for the final implementation!