Calendar Bug Fix
Hey guys, this version of the app has fixed the bug with the calendar
not showing up in full. It now displays 4 weeks at a time. I've also
included a second "project" in this archive file, because it is a part
of the calendar and I wasn't sure if eclipse needs it to compile Globus
properly.