Rep upload of project.
This is an attempt to fix the Caldroid import bug Kelsey was running
into. So, This update includes three files. The first file
(ProjectGlobus_20150225_430) contains two different projects. The Globus
project and the Library project. If this bug persists with this upload,
get rid of this import and import both of the other two files (labeled
with PART 1 and PART 2). If you import both of these as separate
projects, I think it should work just fine.
Let me know if you have any other problems!
-Jesse