Connection working I think the db is down?
the connection is working on the emulator, at least it looks like it is.
Look at the logcat and you can see all of davids messages. Pretty cool
stuff guys. This doesn't send/receive messages yet. I'll get that part
working tonight for create new account. We need to use a thread.
Apparently that makes a difference for some reason. Also, I tried this
on my phone (yeah I got that to work too) and it says it's unable to
connect to the server in the logcat printout. I don't know if the server
is down or if it only works on the emulator. Please Test