Added Caldroid project to documents folder
Ok guys, so this is here to fix our problems loading the calendar
fragment. It looks like the problem was our build path. The trick here
was that when we imported the project from our github uploads, it would
change the location of the source files, giving eclipse a hard attack
whenever it needed to actually use them.
Included with these files is a text file explaining how to get caldroid
set up for you in eclipse so that we don't have to worry about this any
more. Also, after you implement this fix, you should only be exporting
project Globus to github. Do not include the "Library" project.