Updated db interaction
Services are working from activity 1 to 2. Weird errors with db
interaction upon second connection. Takes 3 attempts to connect fully to
the app. We'll have to investigate this. This submission crashes on the
sending to the db the membership and user ID number in order to get the
groups they are a member of back.